Fleshed out TLB shootdown further.

+++ kernel/arc/intr/lapic.c
@@ -129,7 +129,7 @@ void lapic_ack(void)
   lapic_write(LAPIC_EOI, 0);
 }
 
-void lapic_ipi(uint8_t dest, uint8_t mode, uint8_t vector)
+void lapic_ipi(uint8_t dest, uint16_t mode, uint8_t vector)
 {
   /* format the ICR high and low dwords */
   uint32_t icrh = dest << 24;
